move integrations tests to unit tests, because they aren't _really_ integration tests
